SY-4727: fold the dispatch executor into actions.State
