+++ core/pkg/service/lineplot/service.go
@@ -90,7 +90,6 @@ type Service struct {
 	closer io.MultiCloser
 	table  *gorp.Table[Key, LinePlot]
 	state  *actions.State[Key, Action]
-	exec   *actions.Executor[Key, Action]
 }
 
 // OpenService instantiates a new line plot service using the provided configurations.
@@ -101,8 +100,7 @@ func OpenService(ctx context.Context, cfgs ...ServiceConfig) (s *Service, err er
 	if err != nil {
 		return nil, err
 	}
-	s = &Service{cfg: cfg, state: actions.NewState[Key, Action]()}
-	s.exec = actions.NewExecutor(cfg.DB, s.state.Dispatcher())
+	s = &Service{cfg: cfg, state: actions.NewState[Key, Action](cfg.DB)}
 	cleanup, ok := service.NewOpener(ctx, &s.closer)
 	defer func() { err = cleanup(err) }()
 	if s.table, err = gorp.OpenTable(ctx, gorp.TableConfig[Key, LinePlot]{
@@ -166,7 +164,7 @@ func (s *Service) Dispatch(
 	dispatchKey string,
 	acts []Action,
 ) error {
-	return s.exec.Dispatch(ctx, key, dispatchKey, acts, func(tx gorp.Tx) error {
+	return s.state.Dispatch(ctx, key, dispatchKey, acts, func(tx gorp.Tx) error {
 		return s.table.NewUpdate().Where(gorp.MatchKeys[Key, LinePlot](key)).
 			ChangeErr(func(_ gorp.Context, p LinePlot) (LinePlot, error) {
 				return Reduce(p, acts...)
