+++ core/pkg/service/log/service.go
@@ -88,7 +88,6 @@ type Service struct {
 	closer xio.MultiCloser
 	table  *gorp.Table[Key, Log]
 	state  *actions.State[Key, Action]
-	exec   *actions.Executor[Key, Action]
 }
 
 // OpenService instantiates a new log service using the provided configurations. Each
@@ -99,8 +98,7 @@ func OpenService(ctx context.Context, cfgs ...ServiceConfig) (s *Service, err er
 	if err != nil {
 		return nil, err
 	}
-	s = &Service{cfg: cfg, state: actions.NewState[Key, Action]()}
-	s.exec = actions.NewExecutor(cfg.DB, s.state.Dispatcher())
+	s = &Service{cfg: cfg, state: actions.NewState[Key, Action](cfg.DB)}
 	cleanup, ok := service.NewOpener(ctx, &s.closer)
 	defer func() { err = cleanup(err) }()
 	if s.table, err = gorp.OpenTable(ctx, gorp.TableConfig[Key, Log]{
@@ -160,7 +158,7 @@ func (s *Service) Dispatch(
 	dispatchKey string,
 	acts []Action,
 ) error {
-	return s.exec.Dispatch(ctx, key, dispatchKey, acts, func(tx gorp.Tx) error {
+	return s.state.Dispatch(ctx, key, dispatchKey, acts, func(tx gorp.Tx) error {
 		return s.table.NewUpdate().Where(gorp.MatchKeys[Key, Log](key)).
 			ChangeErr(func(_ gorp.Context, l Log) (Log, error) {
 				return Reduce(l, acts...)
