+++ x/go/sync/mutex.go
@@ -7,15 +7,16 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
-// Package lock provides locking primitives beyond the standard library's sync.
-package lock
+// Package sync provides synchronization primitives that extend the standard library's
+// sync package.
+package sync
 
 import "sync"
 
-// Keyed provides a mutex per key, created on first use and discarded once no caller
-// holds or waits for it. The zero value is ready to use. A Keyed must not be copied
-// after first use.
-type Keyed[K comparable] struct {
+// KeyedMutex provides a mutex per key, created on first use and discarded once no
+// caller holds or waits for it. The zero value is ready to use. A KeyedMutex must not
+// be copied after first use.
+type KeyedMutex[K comparable] struct {
 	// mu guards locks.
 	mu    sync.Mutex
 	locks map[K]*keyLock
@@ -30,14 +31,14 @@ type keyLock struct {
 
 // Do runs fn while holding the mutex for key and returns fn's error: calls for the
 // same key run one at a time, calls for different keys run concurrently.
-func (k *Keyed[K]) Do(key K, fn func() error) error {
+func (k *KeyedMutex[K]) Do(key K, fn func() error) error {
 	l := k.acquire(key)
 	l.mu.Lock()
 	defer k.release(key, l)
 	return fn()
 }
 
-func (k *Keyed[K]) acquire(key K) *keyLock {
+func (k *KeyedMutex[K]) acquire(key K) *keyLock {
 	k.mu.Lock()
 	defer k.mu.Unlock()
 	if k.locks == nil {
@@ -52,7 +53,7 @@ func (k *Keyed[K]) acquire(key K) *keyLock {
 	return l
 }
 
-func (k *Keyed[K]) release(key K, l *keyLock) {
+func (k *KeyedMutex[K]) release(key K, l *keyLock) {
 	l.mu.Unlock()
 	k.mu.Lock()
 	defer k.mu.Unlock()
